Add Capture JPEG Demo
- Reorganise into 3 folders "CaptureJPEG", "RecordMP4", "StreamRTSP".
- Modify platform.txt
- Change Audio.cpp to AudioStream.cpp
- Change Video.cpp to Videostream.cpp
- Modify WiFi related drv

+++ Arduino_package/hardware/cores/ambpro2/ard_socket.h
@@ -31,15 +31,15 @@ int get_available(int sock);
 
 //int get_available_v6(int sock);
 
-int recv_data(int sock, const uint8_t *data, uint16_t len, int flag);
+int recv_data(int sock, const uint8_t *data, uint32_t len, int flag);
 
-int send_data(int sock, const uint8_t *data, uint16_t len, int flag);
+int send_data(int sock, const uint8_t *data, uint32_t len, int flag);
 // UDP
-int get_receive(int sock, uint8_t *data, int length, int flag, uint32_t *peer_addr, uint16_t *peer_port);
+int get_receive(int sock, uint8_t *data, uint32_t length, int flag, uint32_t *peer_addr, uint16_t *peer_port);
 
 //int get_receive_v6(int server_fd, void *recv_data, int len, int flags, uint32_t *peer_addr, uint16_t *peer_port);
 
-int sendto_data(int sock, const uint8_t *data, uint16_t len, uint32_t peer_ip, uint16_t peer_port);
+int sendto_data(int sock, const uint8_t *data, uint32_t len, uint32_t peer_ip, uint16_t peer_port);
 
 //int sendto_data_v6(int sock, const void *send_data, size_t len, uint32_t peer_ip, uint16_t peer_port);
