+++ Arduino_package/hardware/libraries/WiFi/src/WiFiClient.h
@@ -10,7 +10,7 @@
 class WiFiClient : public Client {
     public:
         WiFiClient();
-        WiFiClient(uint8_t sock);
+        WiFiClient(int sock);
 
         uint8_t status();
         virtual uint8_t connected();
